[feat] 멀티모듈 프로젝트로 전환 1차본.
