+++ api/src/main/java/com/study/api/exception/handler/GlobalExceptionHandler.java
@@ -1,0 +1,21 @@
+package com.study.api.exception.handler;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import com.study.common.constants.ServerStatusCode;
+import com.study.api.domain.response.ApiResponse;
+
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+ 
+    @ExceptionHandler({Exception.class})
+    protected ResponseEntity handleServerException(Exception e) {
+        return new ResponseEntity(ApiResponse.fail(ServerStatusCode.INTERNAL_SERVER_ERROR), HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+}
