+++ api/src/main/resources/application.properties
@@ -1,0 +1,22 @@
+
+# mariaDB connect
+#spring.datasource.hikari.driver-class-name=org.mariadb.jdbc.Driver
+#spring.datasource.hikari.jdbc-url=jdbc:mariadb://127.0.0.1:3306/STUDY?useUnicode=true&characterEncoding=utf8
+spring.datasource.hikari.driver-class-name=net.sf.log4jdbc.sql.jdbcapi.DriverSpy
+spring.datasource.hikari.jdbc-url=jdbc:log4jdbc:mariadb://127.0.0.1:3306/STUDY?useUnicode=true&characterEncoding=utf8
+spring.datasource.hikari.username=root
+spring.datasource.hikari.password=root
+
+#spring.datasource.driver-class-name=net.sf.log4jdbc.sql.jdbcapi.DriverSpy
+#spring.datasource.url=jdbc:log4jdbc:mariadb://127.0.0.1:3306/STUDY?useUnicode=true&characterEncoding=utf8
+#spring.datasource.username=root
+#spring.datasource.password=root
+
+spring.jpa.open-in-view=false
+spring.jpa.generate-ddl=true
+spring.jpa.show-sql=true
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
+
+# Ehcache \u00EC\u0084\u00A4\u00EC\u00A0\u0095
+spring.cache.ehcache.config=classpath:config/ehcache3.0.xml
