+++ jpa/src/main/java/com/study/jpa/repository/MemberRepository.java
@@ -1,0 +1,8 @@
+package com.study.jpa.repository;
+
+import com.study.jpa.entity.MemberEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<MemberEntity, Long>{
+
+}
