+++ settings.gradle
@@ -6,3 +6,9 @@ pluginManagement {
 	}
 }
 rootProject.name = 'study001'
+include 'common'
+include 'api'
+include 'mybatis'
+include 'jpa'
+
+
