--- src/main/java/com/study/study001/Study001Application.java
@@ -1,15 +1,0 @@
-package com.study.study001;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cache.annotation.EnableCaching;
-
-@EnableCaching
-@SpringBootApplication
-public class Study001Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Study001Application.class, args);
-	}
-
-}
