--- src/main/java/com/study/study001/common/config/MariaDbConfig.java
@@ -1,29 +1,0 @@
-package com.study.study001.common.config;
-
-import javax.sql.DataSource;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-
-import com.zaxxer.hikari.HikariConfig;
-import com.zaxxer.hikari.HikariDataSource;
-
-@Configuration
-@PropertySource("classpath:/application.properties")
-public class MariaDbConfig {
-    
-    @Bean
-    @ConfigurationProperties(prefix = "spring.datasource.hikari")
-    public HikariConfig hikariConfig() {
-        return new HikariConfig();
-    }
-
-    @Bean
-    public DataSource dataSource() throws Exception {
-        DataSource dataSource = new HikariDataSource(hikariConfig());
-        return dataSource;
-    }
-
-}
