--- src/main/java/com/study/study001/jpa/repository/MemberRepository.java
@@ -1,8 +1,0 @@
-package com.study.study001.jpa.repository;
-
-import com.study.study001.jpa.entity.MemberEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<MemberEntity, Long>{
-
-}
