--- src/main/java/com/study/study001/repository/MemberMapper.java
@@ -1,14 +1,0 @@
-package com.study.study001.repository;
-
-import java.util.List;
-
-import org.apache.ibatis.annotations.Mapper;
-
-import com.study.study001.domain.dto.Member;
-
-@Mapper
-public interface MemberMapper {
-    List<Member> getMemberList();
-
-    Member getMember(String mId);
-}
