--- src/main/java/com/study/study001/vo/HealthCheckInfo.java
@@ -1,30 +1,0 @@
-package com.study.study001.vo;
-
-import java.time.LocalDateTime;
-
-import org.apache.commons.lang3.builder.ToStringBuilder;
-import org.apache.commons.lang3.builder.ToStringStyle;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-
-
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class HealthCheckInfo {
-    private String clientIp;
-    private LocalDateTime currentTime;
-    private boolean isHealth;
-
-	@Override
-    public String toString() {
-        return ToStringBuilder.reflectionToString(this, ToStringStyle.JSON_STYLE);
-    }
-}
