--- src/main/resources/application.properties
@@ -1,30 +1,0 @@
-
-# mariaDB connect
-#spring.datasource.hikari.driver-class-name=org.mariadb.jdbc.Driver
-#spring.datasource.hikari.jdbc-url=jdbc:mariadb://127.0.0.1:3306/STUDY?useUnicode=true&characterEncoding=utf8
-spring.datasource.hikari.driver-class-name=net.sf.log4jdbc.sql.jdbcapi.DriverSpy
-spring.datasource.hikari.jdbc-url=jdbc:log4jdbc:mariadb://127.0.0.1:3306/STUDY?useUnicode=true&characterEncoding=utf8
-spring.datasource.hikari.username=root
-spring.datasource.hikari.password=root
-
-#spring.datasource.driver-class-name=net.sf.log4jdbc.sql.jdbcapi.DriverSpy
-#spring.datasource.url=jdbc:log4jdbc:mariadb://127.0.0.1:3306/STUDY?useUnicode=true&characterEncoding=utf8
-#spring.datasource.username=root
-#spring.datasource.password=root
-
-spring.jpa.open-in-view=false
-spring.jpa.generate-ddl=true
-spring.jpa.show-sql=true
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
-# Ehcache \u00EC\u0084\u00A4\u00EC\u00A0\u0095
-spring.cache.ehcache.config=classpath:config/ehcache3.0.xml
-
-# model \u00ED\u0094\u0084\u00EB\u00A1\u009C\u00ED\u008D\u00BC\u00ED\u008B\u00B0 camel case \u00EC\u0084\u00A4\u00EC\u00A0\u0095
-mybatis.configuration.map-underscore-to-camel-case=true
-
-# xml\u00ED\u008C\u008C\u00EC\u009D\u00BC result type\u00EC\u0097\u0090 \u00ED\u008C\u00A8\u00ED\u0082\u00A4\u00EC\u00A7\u0080\u00EB\u00AA\u0085\u00EC\u009D\u0084 \u00EC\u0083\u009D\u00EB\u009E\u00B5\u00ED\u0095\u00A0 \u00EC\u0088\u0098 \u00EC\u009E\u0088\u00EB\u008F\u0084\u00EB\u00A1\u009D alias \u00EC\u0084\u00A4\u00EC\u00A0\u0095
-mybatis.type-aliases-package=com.study.study001.domain
-
-# Mybatis mapper \u00EC\u009C\u0084\u00EC\u00B9\u0098 \u00EC\u0084\u00A4\u00EC\u00A0\u0095
-mybatis.mapper-locations=classpath:mapper/*.xml
