--- src/main/resources/mapper/MemberMapper.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.study.study001.repository.MemberMapper">
-
-    <select id="getMemberList" resultType="Member">
-        SELECT 
-            MNO
-            , MID
-            , MNAME
-            , MTAG
-            , JOIN_DATE
-        FROM 
-            MEMBER_T
-        WHERE
-            1=1            
-    </select>
-
-    <select id="getMember" resultType="Member">
-        SELECT
-            MNO
-             , MID
-             , MNAME
-             , MTAG
-             , JOIN_DATE
-        FROM
-            MEMBER_T
-        WHERE
-            1=1
-            AND MID = #{mId}
-    </select>
-</mapper>
