--- src/test/java/com/study/study001/Study001ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.study.study001;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Study001ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
